# Optimize local database access